modifcacion formato localtimestamp

+++ src/main/java/streaming/objects/Event.java
@@ -11,14 +11,13 @@ public class Event {
     public String userId;
     public String source;
 
-
     @JsonProperty("payload")
-    private void unpackNested(Map<String,Object> payload) {
-        this.url = (String)payload.get("url");
+    private void unpackNested(Map<String, Object> payload) {
+        this.url = (String) payload.get("url");
         this.type = (String) payload.get("type");
-        Double localTimestamp= (Double) payload.get("localTimeStamp");
-        this.localTimestamp= localTimestamp;
-        this.userId=(String) payload.get("userId");
+        Long localTimestamp = (Long) payload.get("localTimeStamp");
+        this.localTimestamp = localTimestamp.doubleValue();
+        this.userId = (String) payload.get("userId");
         this.source = (String) payload.get("source");
     }
 }
