+++ src/main/java/streaming/objects/Query.java
@@ -11,11 +11,11 @@ public class Query {
     public String userId;
 
     @JsonProperty("payload")
-    private void unpackNested(Map<String,Object> payload){
-        this.url=(String)payload.get("url");
-        Double localTimestamp= (Double) payload.get("localTimeStamp");
-        this.localTimestamp= localTimestamp;
-        this.query= (String) payload.get("query");
-        this.userId=(String) payload.get("userId");
+    private void unpackNested(Map<String, Object> payload) {
+        this.url = (String) payload.get("url");
+        Long localTimestamp = (Long) payload.get("localTimeStamp");
+        this.localTimestamp = localTimestamp.doubleValue();
+        this.query = (String) payload.get("query");
+        this.userId = (String) payload.get("userId");
     }
 }
